// src/lib/sanity/types.ts
export type GetWorkoutRecordQueryResult = {
  _id: string;
  _type: "workout";
  _createdAt: string;
  date: string | null;
  duration: number | null;
  exercises: Array<{
    exercise: {
      _id: string;
      name: string | null;
      description: string | null;
    } | null;
    sets: Array<{
      reps: number | null;
      weight: number | null;
      weightUnit: "kg" | "lbs" | null;
      _type: "set";
      _key: string;
    }> | null;
    _type: "workoutExercise";
    _key: string;
  }> | null;
} | null;